Avoid the sleep style for doing nothing
the work load in Syncer.sync may like this before:
loop {
  1, consume all buffer job in jobChan.
  2, sleep 10ms for doing nothing although we are pushing job into jobChan
in another goroutine.
}

this hurts especial when the buffer size of jobChan is small.
most of the time will just sleeping.

This PR avoid the bad sleep code style, and just keep the same logic.
Note for dest=file/kafka will never `execute` in the maxExecutionWaitTime case